requirements.txt
# Python version: 3.9-3.10 recommended for Streamlit Cloud
# Python 3.11+ may cause compatibility issues with Whisper

streamlit>=1.52.1
openai-whisper>=20231117
librosa>=0.10.1
# pydub>=0.25.1  # Không cần nữa - đã thay bằng librosa/soundfile để tránh phụ thuộc ffprobe
soundfile>=0.12.1
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
python-docx>=1.1.0
reportlab>=4.0.0
scipy>=1.11.0
torch>=2.0.0
torchaudio>=2.0.0
imageio-ffmpeg>=0.5.0
sentencepiece>=0.1.99
protobuf>=3.20.0,<5.0.0
transformers>=4.30.0
accelerate>=0.20.0
jiwer>=3.0.0
sacrebleu>=2.0.0
langdetect>=1.0.9
faster-whisper>=1.0.0
# tf-keras>=2.15.0  # Not required for Whisper
# tensorflow>=2.15.0  # Optional - chỉ cần nếu muốn dùng tf-keras
psutil>=5.9.0  # For memory monitoring

# Dependencies cho Whisper:
# - openai-whisper: Whisper model
# - torch, torchaudio: PyTorch backend
# - imageio-ffmpeg: FFmpeg cho Whisper

# API/Streaming
fastapi>=0.110.0
uvicorn>=0.23.0
audio-recorder-streamlit>=0.0.8
altair>=5.0.0

# Gemini AI (Optional - for enhanced text improvement)
google-generativeai>=0.3.0
